fix: fixing small issues
