+++ README.md
@@ -17,6 +17,8 @@ Code implementation for our robot competing in the EPFL Robotic Competition (MAK
 
 ### 1. Nvidia Jetson Nano Connection
 
+#### 1.1 Wireless Connection
+
 1. Connect to the zerotier network (ID: `856127940c82f9a4`):
 ```
 sudo zerotier-cli join 856127940c82f9a4
@@ -31,6 +33,16 @@ psswd: jetson
 3. Shutdown Jetson Nano:
 ```
 sudo poweroff
+```
+
+#### 1.2 Wired Connection (Ethernet)
+
+1. Connect the Jetson Nano to the laptop with an Ethernet cable.
+
+2. SSH into the Jetson Nano:
+```
+ssh jetson@192.168.1.100
+psswd: jetson
 ```
 
 ### 2. Docker Container Setup (Run ROS2 on Jetson Nano)
